Use __init__ instead of __new__ for Bot constructors
Pylance was complaining about no overload found whenever the Bot
class would be inherited and provided their own constructor.

I noticed the stubs used `__new__` instead of `__init__` so I
changed it to reflect the actual code in discord.py (which does not
use `__new__`). I tested this in both Pylance and Mypy and it seems
to work without issues.